PBL - 2/requirement.txt
qiskit
qiskit-aer
pycryptodome
numpy
matplotlib